Fix F to allow for F()
It's much easier to compose an F incrementally if you can start with
an empty F() that behaves correctly. This fixes F and F usage so that
it accounts for empty F()s.

e.g.:

    F() | F(tag="awesome")

Prior to this change, that would produce:

    {"or_": [{}, {"tag": "awesome"}]}

Which makes ES grumpy. Now it produces:

    {"tag": "awesome"}

because the empty F() gets ignored.