Issue 38. Explain support
This adds support for 'explain' flag in searches, documentation, tests,
and a function that formats the resulting explanation output so it's
human-readable.

The latter is kind of a nice-to-have, so I stuck it in utils.py. It's
not documented, yet, because I figure it could use some work.

Conflicts:

	docs/debugging.rst
	docs/queries.rst
	elasticutils/__init__.py
	elasticutils/tests/test_query.py

I nixed the docs and the tests---this all works in the master branch, so
I'm reasonably sure it works in the sumo branch.